% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownfield_nlud_shp.R
\docType{data}
\name{brownfield_nlud_shp}
\alias{brownfield_nlud_shp}
\title{Brownfield NLUD Shapefile 2009-2010}
\format{A shapefile with 1885 features and 15 fields:
\describe{
  \item{REGIONNAME}{region name}
  \item{COUNTY}{county}
  \item{LA_CODE}{local authority code}
  \item{LA_NAME}{local authority name}
  \item{SITE_REF}{site reference}
  \item{EASTING}{easting}
  \item{NORTHING}{northing}
  \item{AREA}{area, in hectares}
  \item{SAO}{site info}
  \item{PAO}{site info}
  \item{STREET}{street}
  \item{locality}{locality}
  \item{TOWN}{town}
  \item{ADMIN}{admin}
  \item{POSTCODE}{postcode}
  ...
}}
\source{
\url{http://www.data.london.gov.uk/dataset/london-brownfield-sites-review/}
}
\usage{
brownfield_nlud_shp
}
\description{
A dataset containing the brownfield polygons from the National Land Use Database 2009-2010.
brownfield_nlud_shp.
}
\keyword{datasets}
