% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownfieldgrid_nlud_satellite.R
\name{brownfieldgrid_nlud_satellite}
\alias{brownfieldgrid_nlud_satellite}
\title{Interactive Grid Maps of NLUD Brownfield Land}
\usage{
brownfieldgrid_nlud_satellite(location)
}
\arguments{
\item{Location}{A location in London}
}
\value{
Interactive grid map of NLUD brownfield land.
}
\description{
Function for creating interactive grid maps of NLUD brownfield land.
}
\examples{
brownfieldgrid_nlud_satellite("Limehouse, London")
}
