% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownfieldgrid_osm_borough.R
\name{brownfieldgrid_osm_borough}
\alias{brownfieldgrid_osm_borough}
\title{Grid Maps of Brownfield Land and Borough}
\usage{
brownfieldgrid_osm_borough(location)
}
\arguments{
\item{Location}{A location in London}
}
\value{
Grid map of brownfield land and borough outline.
}
\description{
Function for creating grid maps of OSM brownfield land and the borough outline.
}
\examples{
brownfieldgrid_osm_borough("Limehouse, London")
}
