% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownfield_osm_borough_nogrid.R
\name{brownfieldgrid_osm_borough_nogrid}
\alias{brownfieldgrid_osm_borough_nogrid}
\title{OSM Brownfield Land by Borough}
\usage{
brownfieldgrid_osm_borough_nogrid(borough)
}
\arguments{
\item{Location}{A London borough}
}
\value{
Map of brownfield land by borough.
}
\description{
Function for creating maps of OSM brownfield land by borough.
}
