% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownfield_osm_londondata.R
\name{brownfieldgrid_osm_londondata}
\alias{brownfieldgrid_osm_londondata}
\title{Download OSM Brownfield Land Data for Greater London}
\usage{
brownfieldgrid_osm_londondata()
}
\value{
OSM brownfield land use data
}
\description{
Function for downloading OSM brownfield land use data.
}
\examples{
brownfieldgrid_osm_londondata()
}
