% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownfieldgrid_osm_nlud.R
\name{brownfieldgrid_osm_nlud}
\alias{brownfieldgrid_osm_nlud}
\title{Grid Maps of OSM and NLUD Brownfield Land}
\usage{
brownfieldgrid_osm_nlud(location)
}
\arguments{
\item{Location}{A location in London}
}
\value{
Grid map of OSM and NLUD brownfield land.
}
\description{
Function for creating grid maps of OSM and NLUD brownfield land for comparison.
}
\examples{
brownfieldgrid_osm_nlud("Limehouse, London")
}
