% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownfieldgrid_osm_satellite.R
\name{brownfieldgrid_osm_satellite}
\alias{brownfieldgrid_osm_satellite}
\title{Interactive Grid Maps of Brownfield Land and Borough}
\usage{
brownfieldgrid_osm_satellite(location)
}
\arguments{
\item{Location}{A location in London}
}
\value{
Interactive grid map of brownfield land and borough outline.
}
\description{
Function for creating interactive grid maps of OSM brownfield land.
}
\examples{
brownfieldgrid_osm_satellite("Limehouse, London")
}
