% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownfield_register_nlud.R
\name{brownfieldgrid_register_nlud}
\alias{brownfieldgrid_register_nlud}
\title{Grid Maps of Brownfield Register 2018 and NLUD 2009-2010 Brownfields}
\usage{
brownfieldgrid_register_nlud(location)
}
\arguments{
\item{Location}{A location in London}
}
\value{
Grid map of Brownfield Register and NLUD brownfields.
}
\description{
Function for creating grid maps of Brownfield Register 2018 and NLUD 2009-2010 brownfields for comparison.
}
\examples{
brownfieldgrid_register_nlud("Limehouse, London")
}
