% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london_boro_shp.R
\docType{data}
\name{london_boro_shp}
\alias{london_boro_shp}
\title{London Borough Shapefile}
\format{A shapefile with 33 features and 7 fields:
\describe{
  \item{NAME}{borough name}
  \item{GSS_CODE}{gss code}
  \item{HECTARES}{borough size, hectares}
  \item{NONLD_AREA}{nold area}
  \item{ONS_INNER}{osn inner, y or n}
  \item{SUB_2009}{sub 2009}
  \item{SUB_2006}{sub 2006}
  ...
}}
\source{
\url{https://data.london.gov.uk/dataset/statistical-gis-boundary-files-london/}
}
\usage{
london_boro_shp
}
\description{
A dataset containing the London borough polygons (exclusing MHW).
london_boro_shp.
}
\keyword{datasets}
